+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.6.3
 
-  * 
+  * Fix started? and started methods for badges with a combination of custom and normal requirements
 
 ## Version 0.6.2